day 6 part 2 rust solution (brute force)

+++ 2023/src/day6/boat_race.rs
@@ -2,11 +2,11 @@
 #[derive(Debug, Clone, Copy)]
 pub struct BoatRace {
   pub id: u16,
-  pub time: u16,
-  pub distance: u16,
+  pub time: u64,
+  pub distance: u64,
 }
 impl BoatRace {
-  pub fn hold_button(&self, ms: u16) -> u16 {
+  pub fn hold_button(&self, ms: u64) -> u64 {
     let remaining_time = self.time - ms;
     let distance = remaining_time * ms;
     distance
